move SendOverview to components
