+++ frontend/src/components/GddSend/QrCode.spec.js
