+++ frontend/src/components/GddSend/QrCode.vue
